Move code to its own file
